on edit functionality for Enquiry and detailEnquiry - Subcomponents for student details
